requirements.txt
python-dotenv
minify-html
openpyxl
pandas
tabulate
xlrd